step2 refactor - 피드백 내용 반영
