+++ src/main/java/step2/car/domain/engine/RandomEngine.java
@@ -1,4 +1,4 @@
-package step2.car.domain.car.engine;
+package step2.car.domain.engine;
 
 import java.util.Random;
