+++ src/test/java/step2/race/domain/RaceTest.java
@@ -1,9 +1,9 @@
 package step2.race.domain;
 
 import org.junit.jupiter.api.Test;
-import step2.car.domain.car.Car;
-import step2.car.domain.car.engine.Engine;
-import step2.car.domain.car.engine.FixedEngine;
+import step2.car.domain.Car;
+import step2.car.domain.engine.Engine;
+import step2.car.domain.engine.FixedEngine;
 
 import java.util.Arrays;
